+++ app/test/storage_migration/golden_harness.dart
@@ -220,6 +220,7 @@ Future<GoldenBuild> buildGoldenConfig(StorageSandbox box) async {
     vpnMode: await SettingsStorage.getVpnMode(),
     idleSuspend: await SettingsStorage.getIdleSuspend(),
     idleSuspendReachable: await SettingsStorage.getIdleSuspendReachable(),
+    wgBuildMax: await SettingsStorage.getWgBuildMax(),
     passiveCheck: await SettingsStorage.getPassiveCheck(),
     tailscaleStateRoot: kGoldenTailscaleStateRoot,
   );